Move control4 base entity to separate module
